Creating a package intended for helpers and the PrivacyHelper class. Changing privacy rules (DRL file) in order to use the created helper.
